+++ mdw/mdw.go
@@ -1,0 +1,114 @@
+package mdw
+// setting aside our middleware
+
+import (
+	"net/http"
+	"log"
+	"time"
+	"reflect"
+	"encoding/json"
+
+	"github.com/gorilla/context"
+
+	"github.com/sh4t/community/errs"
+)
+
+
+func RecoverHandler(next http.Handler) http.Handler {
+	fn := func(w http.ResponseWriter, r *http.Request) {
+		defer func() {
+			if err := recover(); err != nil {
+				log.Printf("panic: %+v", err)
+				errs.WriteError(w, errs.ErrInternalServer)
+			}
+		}()
+
+		next.ServeHTTP(w, r)
+	}
+
+	return http.HandlerFunc(fn)
+}
+
+// cors is a bit annoying, but here is an attempt to resolve for community-ui angular love.
+func CorsHandler(next http.Handler) http.Handler {
+	fn := func(w http.ResponseWriter, r *http.Request) {
+
+		if origin := r.Header.Get("Origin"); origin != "" {
+		    w.Header().Set("Access-Control-Allow-Origin", origin)
+		    w.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
+		    w.Header().Set("Access-Control-Allow-Headers", "Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization")
+		}
+		
+	    if r.Method == "OPTIONS" {
+	        return
+	    }
+
+	    next.ServeHTTP(w, r)
+	}
+
+	return http.HandlerFunc(fn)
+}
+
+func LoggingHandler(next http.Handler) http.Handler {
+	fn := func(w http.ResponseWriter, r *http.Request) {
+		t1 := time.Now()
+		next.ServeHTTP(w, r)
+		t2 := time.Now()
+		log.Printf("[%s] %q %v\n", r.Method, r.URL.String(), t2.Sub(t1))
+	}
+
+	return http.HandlerFunc(fn)
+}
+
+func AcceptHandler(next http.Handler) http.Handler {
+	fn := func(w http.ResponseWriter, r *http.Request) {
+		if r.Header.Get("Accept") != "application/vnd.api+json" {
+			errs.WriteError(w, errs.ErrNotAcceptable)
+			return
+		}
+
+		next.ServeHTTP(w, r)
+	}
+
+	return http.HandlerFunc(fn)
+}
+
+func ContentTypeHandler(next http.Handler) http.Handler {
+	fn := func(w http.ResponseWriter, r *http.Request) {
+		if r.Header.Get("Content-Type") != "application/vnd.api+json" {
+			errs.WriteError(w, errs.ErrUnsupportedMediaType)
+			return
+		}
+
+		next.ServeHTTP(w, r)
+	}
+
+	return http.HandlerFunc(fn)
+}
+
+
+
+func BodyHandler(v interface{}) func(http.Handler) http.Handler {
+	t := reflect.TypeOf(v)
+
+	m := func(next http.Handler) http.Handler {
+		fn := func(w http.ResponseWriter, r *http.Request) {
+			val := reflect.New(t).Interface()
+			err := json.NewDecoder(r.Body).Decode(val)
+
+			if err != nil {
+				errs.WriteError(w, errs.ErrBadRequest)
+				return
+			}
+
+			if next != nil {
+				context.Set(r, "body", val)
+				next.ServeHTTP(w, r)
+			}
+		}
+
+		return http.HandlerFunc(fn)
+	}
+
+	return m
+}
